Command classes that follow the COMMAND_CLASS_MARK in the node info list (i.e. those that the device can control in other devices) are now created, but without values.
This allows us to create the Basic class for devices such as sensors that do not respond to basic commands (but do send them to the controller to signal an event has occured), without having a value show up in the UI.

BasicSet commands now send a notification when received.  The notification is Type_NodeEvent, renamed from the old Type_NodeStatus for clarity of purpose.  The notification contains the value sent in the BasicSet command, so users can tell whether the event has started (non-zero) or stopped (zero)